pyproject.toml
[project]
name = "deep-agents-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.1",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "tavily-python>=0.7.17",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"
